feat(metastore/postgres): split test files
